Major refactoring of components and files | results configured
