+++ src/app/components/quiz-shows/components/quiz-show-card/quiz-show-card.component.spec.ts
