--- src/app/components/quiz-shows/quiz-result/quiz-result.component.html
@@ -1,1 +1,0 @@
-<p>quiz-result works!</p>
