--- src/app/components/quiz-shows/quiz-result/quiz-result.component.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { QuizResultComponent } from './quiz-result.component';
-
-describe('QuizResultComponent', () => {
-  let component: QuizResultComponent;
-  let fixture: ComponentFixture<QuizResultComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ QuizResultComponent ]
-    })
-    .compileComponents();
-
-    fixture = TestBed.createComponent(QuizResultComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
