--- src/app/components/quiz-shows/quiz-result/quiz-result.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'bs-quiz-result',
-  templateUrl: './quiz-result.component.html',
-  styleUrls: ['./quiz-result.component.scss']
-})
-export class QuizResultComponent {
-
-}
