+++ src/app/components/quiz-shows/quiz-shows.component.ts
@@ -19,7 +19,6 @@ export class QuizShowsComponent implements OnInit {
     this.getDataAndSetInStorage();
   }
 
-  //? BRIDGE POSSIBLE
   private getDataAndSetInStorage(): void {
     const showData = this.storageSVC.getShows();
     if (showData) {
