--- src/app/components/quiz-shows/shared/quiz-season-card/quiz-season-card.component.html
@@ -1,1 +1,0 @@
-<p>quiz-season-card works!</p>
