--- src/app/components/quiz-shows/shared/quiz-season-card/quiz-season-card.component.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { QuizSeasonCardComponent } from './quiz-season-card.component';
-
-describe('QuizSeasonCardComponent', () => {
-  let component: QuizSeasonCardComponent;
-  let fixture: ComponentFixture<QuizSeasonCardComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [ QuizSeasonCardComponent ]
-    })
-    .compileComponents();
-
-    fixture = TestBed.createComponent(QuizSeasonCardComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
