--- src/app/components/quiz-shows/shared/quiz-season-card/quiz-season-card.component.ts
@@ -1,13 +1,0 @@
-import { Component } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-@Component({
-  selector: 'bs-quiz-season-card',
-  standalone: true,
-  imports: [CommonModule],
-  templateUrl: './quiz-season-card.component.html',
-  styleUrls: ['./quiz-season-card.component.scss']
-})
-export class QuizSeasonCardComponent {
-
-}
