--- src/app/core/shared/interfaces/core.enums.ts
@@ -1,9 +1,0 @@
-enum Paths {
-  STATIC = 'info',
-  AUTH = 'auth',
-  HOME = 'shows',
-  PROFILE = 'profile'
-}
-
-
-export { Paths };
