+++ src/app/models/core.model.ts
