+++ src/app/models/shared/global.models.ts
@@ -1,2 +1,10 @@
 export type GetParam = string | null
+
 export type QuizButton = 'Next' | 'Finish' | 'Begin';
+
+export enum Paths {
+  HOME = 'shows',
+  STATIC = 'info',
+  AUTH = 'auth',
+  PROFILE = 'profile'
+}
