+++ src/app/shared/services/storage.service.ts
@@ -11,6 +11,10 @@ export class StorageService {
     localStorage.setItem(StorageKeys.IN_PROGRESS, 'true');
   }
 
+  removeQuizInit(): void {
+    this.removeItem(StorageKeys.IN_PROGRESS);
+  }
+
   setShows(value: ShowWithId[]): void {
     localStorage.setItem(StorageKeys.SHOWS, this.toJSON(value));
   }
